Rename file fetchErrorHandler
