+++ lib/pages/home_page.dart
@@ -10,7 +10,7 @@ import 'package:trufi_core/blocs/home_page_cubit.dart';
 import 'package:trufi_core/blocs/preferences_cubit.dart';
 import 'package:trufi_core/l10n/trufi_localization.dart';
 import 'package:trufi_core/models/map_route_state.dart';
-import 'package:trufi_core/widgets/fetchErrorHandler.dart';
+import 'package:trufi_core/widgets/fetch_error_handler.dart';
 import 'package:trufi_core/widgets/from_marker.dart';
 import 'package:trufi_core/widgets/to_marker.dart';
